modify sql file

+++ peerprep/src/main/resources/sql/peerprep_interview_tables.sql
@@ -1,28 +1,28 @@
 CREATE TABLE IF NOT EXISTS `interview_questions` (
-  `id` bigint PRIMARY KEY AUTO_INCREMENT,
+  `id` int unsigned PRIMARY KEY AUTO_INCREMENT,
   `title` varchar(255),
   `content` text,
   `difficulty` int
 );
 
 CREATE TABLE IF NOT EXISTS `interview_solutions` (
-  `id` bigint PRIMARY KEY AUTO_INCREMENT,
-  `question_id` bigint,
+  `id` int unsigned PRIMARY KEY AUTO_INCREMENT,
+  `question_id` int unsigned,
   `content` text
 );
 
 CREATE TABLE IF NOT EXISTS `users_questions_history` (
-  `id` bigint PRIMARY KEY AUTO_INCREMENT,
-  `user_id` bigint,
-  `question_id` bigint,
+  `id` int unsigned PRIMARY KEY AUTO_INCREMENT,
+  `user_id` int unsigned,
+  `question_id` int unsigned,
   `user_answer` text,
   `created_at` timestamp,
   `deleted_at` timestamp
 );
 
 CREATE TABLE IF NOT EXISTS `users_roles_history` (
-  `id` bigint PRIMARY KEY AUTO_INCREMENT,
-  `user_id` bigint,
+  `id` int unsigned PRIMARY KEY AUTO_INCREMENT,
+  `user_id` int unsigned,
   `role` int,
   `created_at` timestamp,
   `deleted_at` timestamp
